+++ src/main/java/fuzs/goldenagecombat/client/handler/ClientCooldownHandler.java
@@ -20,7 +20,7 @@ public class ClientCooldownHandler {
             // just for that reason the whole indicator is also disabled later on
             ClassicCombatHandler.disableCooldownPeriod(this.minecraft.player);
         }
-        if (GoldenAgeCombat.CONFIG.client().classic.hideAttackIndicator) {
+        if (GoldenAgeCombat.CONFIG.server().classic.removeCooldown) {
             // indicator would otherwise render when looking at an entity, even when there is no cooldown
             this.attackIndicator = this.minecraft.options.attackIndicator;
             this.minecraft.options.attackIndicator = AttackIndicatorStatus.OFF;
@@ -30,7 +30,7 @@ public class ClientCooldownHandler {
     @SubscribeEvent
     public void onRenderGameOverlay$Post(final RenderGameOverlayEvent.Post evt) {
         if (evt.getType() != RenderGameOverlayEvent.ElementType.ALL) return;
-        if (GoldenAgeCombat.CONFIG.client().classic.hideAttackIndicator) {
+        if (GoldenAgeCombat.CONFIG.server().classic.removeCooldown) {
             // reset to old value
             // we don't just leave this disabled as it'll change the vanilla setting permanently, which a mod shouldn't do imo
             this.minecraft.options.attackIndicator = this.attackIndicator;