+++ src/main/java/fuzs/goldenagecombat/client/handler/SwordBlockingRenderer.java
@@ -28,7 +28,7 @@ public class SwordBlockingRenderer {
             HumanoidArm handSide = isMainHand ? player.getMainArm() : player.getMainArm().getOpposite();
             boolean isHandSideRight = handSide == HumanoidArm.RIGHT;
             ((ItemInHandRendererAccessor) itemRenderer).callApplyItemArmTransform(matrixStack, handSide, evt.getEquipProgress());
-            if (GoldenAgeCombat.CONFIG.client().animations.attackWhileUsing) {
+            if (GoldenAgeCombat.CONFIG.server().classic.attackWhileUsing) {
                 ((ItemInHandRendererAccessor) itemRenderer).callApplyItemArmAttackTransform(matrixStack, handSide, evt.getSwingProgress());
             }
             this.transformBlockFirstPerson(matrixStack, handSide);