+++ src/main/java/fuzs/goldenagecombat/mixin/FoodDataMixin.java
@@ -33,8 +33,12 @@ public abstract class FoodDataMixin {
                 this.tickCombatTest(player);
                 callbackInfo.cancel();
             }
-            case LEGACY -> {
+            case LEGACY_COMBAT -> {
                 this.tickLegacy(player);
+                callbackInfo.cancel();
+            }
+            case CUSTOM -> {
+                this.tickCustom(player);
                 callbackInfo.cancel();
             }
         }
@@ -101,6 +105,39 @@ public abstract class FoodDataMixin {
         } else if (this.foodLevel <= 0) {
             ++this.tickTimer;
             if (this.tickTimer >= 80) {
+                if (player.getHealth() > 10.0F || difficulty == Difficulty.HARD || player.getHealth() > 1.0F && difficulty == Difficulty.NORMAL) {
+                    player.hurt(DamageSource.STARVE, 1.0F);
+                }
+                this.tickTimer = 0;
+            }
+        } else {
+            this.tickTimer = 0;
+        }
+    }
+
+    @Unique
+    private void tickCustom(Player player) {
+        Difficulty difficulty = player.level.getDifficulty();
+        this.lastFoodLevel = this.foodLevel;
+        if (this.exhaustionLevel > 4.0F) {
+            this.exhaustionLevel -= 4.0F;
+            if (this.saturationLevel > 0.0F) {
+                this.saturationLevel = Math.max(this.saturationLevel - 1.0F, 0.0F);
+            } else if (difficulty != Difficulty.PEACEFUL) {
+                this.foodLevel = Math.max(this.foodLevel - 1, 0);
+            }
+        }
+        boolean flag = player.level.getGameRules().getBoolean(GameRules.RULE_NATURAL_REGENERATION);
+        if (flag && this.foodLevel > 6 && player.isHurt()) {
+            ++this.tickTimer;
+            if (this.tickTimer >= 60) {
+                player.heal(1.0F);
+                this.addExhaustion(6.0F);
+                this.tickTimer = 0;
+            }
+        } else if (this.foodLevel <= 0) {
+            ++this.tickTimer;
+            if (this.tickTimer >= 60) {
                 if (player.getHealth() > 10.0F || difficulty == Difficulty.HARD || player.getHealth() > 1.0F && difficulty == Difficulty.NORMAL) {
                     player.hurt(DamageSource.STARVE, 1.0F);
                 }