+++ src/main/java/fuzs/goldenagecombat/mixin/ProjectileUtilMixin.java
@@ -1,5 +1,6 @@
 package fuzs.goldenagecombat.mixin;
 
+import fuzs.goldenagecombat.GoldenAgeCombat;
 import net.minecraft.world.entity.projectile.ProjectileUtil;
 import net.minecraft.world.phys.AABB;
 import org.spongepowered.asm.mixin.Mixin;
@@ -10,10 +11,11 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 public abstract class ProjectileUtilMixin {
     @ModifyVariable(method = "getEntityHitResult(Lnet/minecraft/world/entity/Entity;Lnet/minecraft/world/phys/Vec3;Lnet/minecraft/world/phys/Vec3;Lnet/minecraft/world/phys/AABB;Ljava/util/function/Predicate;D)Lnet/minecraft/world/phys/EntityHitResult;", at = @At(value = "STORE"), ordinal = 1)
     private static AABB getEntityHitResult$aabb(AABB oldAABB) {
-        // TODO add config option
-        final double minSize = 0.9F;
-        if (oldAABB.getXsize() < minSize || oldAABB.getYsize() < minSize || oldAABB.getZsize() < minSize) {
-            return oldAABB.inflate(Math.max(0.0, minSize - oldAABB.getXsize()) / 2.0, Math.max(0.0, minSize - oldAABB.getYsize()) / 2.0, Math.max(0.0, minSize - oldAABB.getZsize()) / 2.0);
+        if (GoldenAgeCombat.CONFIG.server().adjustments.minHitboxSize) {
+            final double minSize = 0.9F;
+            if (oldAABB.getXsize() < minSize || oldAABB.getYsize() < minSize || oldAABB.getZsize() < minSize) {
+                return oldAABB.inflate(Math.max(0.0, minSize - oldAABB.getXsize()) / 2.0, Math.max(0.0, minSize - oldAABB.getYsize()) / 2.0, Math.max(0.0, minSize - oldAABB.getZsize()) / 2.0);
+            }
         }
         return oldAABB;
     }